requirements.txt
Django>=5.0,<6
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-redis>=5.4
psycopg[binary]>=3.1
celery[redis]>=5.3
redis>=5.0
httpx>=0.27
python-dotenv>=1.0
gunicorn>=22.0
